@ P96 *Info layout alignment with RadeonGCN-OS4 + cursor signed-coord fix
Aligned the P96 chip-driver headers to the canonical Picasso96 SDK /
RadeonGCN-OS4 layout, and fixed a cursor coordinate bug surfaced by it.

boardinfo.h:
- struct BitMapExtra rebuilt to the canonical 46-byte layout (MinNode
  BoardNode + separate HashChain) -- the old reconstruction conflated
  them, shifting every field from Match onward 4 bytes early.  Added
  the BMEF_* flag defines.
- Slots 58/59/60 (AllocBitMap/FreeBitMap/GetBitMapAttr) given the
  canonical signatures; added a struct TagItem forward declaration.
- RGBFormats: WORD -> UWORD (canonical).
- ChipLine tail: replaced the spurious LinePtrnTotal with the canonical
  pad/Xorigin/Yorigin (38-byte struct).
- SetSpritePosition (slot 62): UWORD x,y -> WORD x,y.  graphics.library
  passes signed coordinates when the pointer parks off the top/left
  edge; the unsigned declaration mis-handled the sign and propagated a
  ~4-billion value into the VirtIO cursor command.

chip_alloc.c:
- An ABI probe confirmed AOS4 rtg.library never invokes slots 58/59/60;
  removed the non-canonical AllocBitMap/FreeBitMap/GetBitMapAttr
  implementations and left the slots NULL (graphics.library fallback).

chip_gpu_cmds.c:
- chip_CursorMove takes WORD; chip_CursorMove and chip_CursorUpdate
  clamp negative cursor positions to 0 before sending the unsigned
  VirtIO GPU cursor command.

@

+++ include/chip/chip_state.h
@@ -845,7 +845,7 @@ BOOL chip_GetEDID(struct ChipGPUState *gs, uint32 scanout);
 void chip_handle_display_change(struct ChipGPUState *gs);
 BOOL chip_CursorCreate(struct ChipGPUState *gs);
 void chip_CursorUpdate(struct ChipGPUState *gs, uint32 hot_x, uint32 hot_y);
-void chip_CursorMove(struct ChipGPUState *gs, uint32 x, uint32 y);
+void chip_CursorMove(struct ChipGPUState *gs, WORD x, WORD y);
 void chip_CursorHide(struct ChipGPUState *gs);
 
 /* chip_gpu_3d.c */
